refactor: organize and modularize code
