+++ src/games/sound-quiz/components/index.ts
@@ -1,0 +1,5 @@
+export * from "./actionButton";
+export * from "./header";
+export * from "./levelControl";
+export * from "./options";
+export * from "./timer";
