+++ src/games/sound-quiz/states/interfaces.ts
